DJANGO/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
distlib==0.3.6
Django==3.2.13
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
gunicorn==20.1.0
idna==3.4
mysql==0.0.3
mysqlclient==2.1.1
Pillow==9.4.0
platformdirs==2.6.2
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
virtualenv==20.17.1
